requirements.txt
streamlit
python-dotenv
youtube-transcript-api
langchain
openai
tiktoken
langchain_community
langchain_openai